requirements.txt
langchain>=0.0.267
faiss-cpu>=1.7.4
openai>=0.27.0
numpy>=1.23.5
tqdm>=4.65.0
langchain-community
langchain-openai